requirements.txt
aiogram>=3.4
aiohttp>=3.9
SQLAlchemy>=2.0
alembic>=1.13
python-dotenv>=1.0
APScheduler>=3.10
yookassa>=3.0

